pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "api-tester-mcp"
version = "1.0.0"
description = "MCP server for API testing with Swagger/OpenAPI and Postman collection support"
authors = [
    {name = "API Tester MCP", email = "api-tester@example.com"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
keywords = ["mcp", "api-testing", "swagger", "openapi", "postman", "qa", "sdet", "testing"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Testing",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "fastmcp>=0.2.0",
    "pydantic>=2.0.0",
    "requests>=2.28.0",
    "jsonschema>=4.0.0",
    "pyyaml>=6.0",
    "jinja2>=3.1.0",
    "aiofiles>=23.0.0",
    "aiohttp>=3.8.0",
    "faker>=19.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
]

[project.scripts]
api-tester-mcp = "api_tester_mcp.server:main"

[project.urls]
Homepage = "https://github.com/kirti676/api_tester_mcp"
Repository = "https://github.com/kirti676/api_tester_mcp.git"
Issues = "https://github.com/kirti676/api_tester_mcp/issues"
Changelog = "https://github.com/kirti676/api_tester_mcp/blob/main/CHANGELOG.md"

[tool.setuptools]
packages = ["api_tester_mcp"]

[tool.setuptools.package-data]
api_tester_mcp = ["*.py"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
